+++ src/packages/babel.js
@@ -1,60 +1,60 @@
 module.exports = {
-  base: [
+  "base": [
     {
-      '@babel/cli': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/cli": {
+        "version": "7.16.8",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/core': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/core": {
+        "version": "7.16.7",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/preset-env': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/preset-env": {
+        "version": "7.16.8",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/runtime': {
-        version: '7.16.3',
-        type: '--save-dev',
-      },
+      "@babel/runtime": {
+        "version": "7.16.7",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/plugin-proposal-throw-expressions': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/plugin-proposal-throw-expressions": {
+        "version": "7.16.7",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/plugin-transform-arrow-functions': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/plugin-transform-arrow-functions": {
+        "version": "7.16.7",
+        "type": "--save-dev"
+      }
     },
     {
-      '@babel/plugin-transform-runtime': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
+      "@babel/plugin-transform-runtime": {
+        "version": "7.16.8",
+        "type": "--save-dev"
+      }
     },
     {
-      'babel-loader': {
-        version: '8.2.3',
-        type: '--save-dev',
-      },
-    },
+      "babel-loader": {
+        "version": "8.2.3",
+        "type": "--save-dev"
+      }
+    }
   ],
-  react: [
+  "react": [
     {
-      '@babel/preset-react': {
-        version: '7.16.0',
-        type: '--save-dev',
-      },
-    },
-  ],
+      "@babel/preset-react": {
+        "version": "7.16.7",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }