+++ src/packages/esProjectMaker.js
@@ -1,10 +1,10 @@
 module.exports = {
-  base: [
+  "base": [
     {
-      '@mparulski/es-project-maker': {
-        version: '0.28.0',
-        type: '--save-dev',
-      },
-    },
-  ],
+      "@mparulski/es-project-maker": {
+        "version": "0.29.0",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }