+++ src/packages/prettier.js
@@ -1,10 +1,10 @@
 module.exports = {
-  base: [
+  "base": [
     {
-      prettier: {
-        version: '2.4.1',
-        type: '--save-dev',
-      },
-    },
-  ],
+      "prettier": {
+        "version": "2.5.1",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }