+++ src/packages/tests.js
@@ -1,35 +1,35 @@
 module.exports = {
-  base: [
+  "base": [
     {
-      jest: {
-        version: '27.3.1',
-        type: '--save-dev',
-      },
+      "jest": {
+        "version": "27.4.7",
+        "type": "--save-dev"
+      }
     },
     {
-      'babel-jest': {
-        version: '27.3.1',
-        type: '--save-dev',
-      },
+      "babel-jest": {
+        "version": "27.4.6",
+        "type": "--save-dev"
+      }
     },
     {
-      'ts-jest': {
-        version: '27.0.7',
-        type: '--save-dev',
-      },
+      "ts-jest": {
+        "version": "27.1.3",
+        "type": "--save-dev"
+      }
     },
     {
-      '@types/jest': {
-        version: '27.0.2',
-        type: '--save-dev',
-      },
+      "@types/jest": {
+        "version": "27.4.0",
+        "type": "--save-dev"
+      }
     },
     {
-      '@testing-library/react': {
-        version: '12.1.2',
-        type: '--save-dev',
-      },
-    },
+      "@testing-library/react": {
+        "version": "12.1.2",
+        "type": "--save-dev"
+      }
+    }
   ],
-  react: [],
+  "react": []
 }