+++ src/packages/webpack.js
@@ -1,46 +1,46 @@
 module.exports = {
-  base: [
+  "base": [
     {
-      webpack: {
-        version: '5.64.1',
-        type: '--save-dev',
-      },
+      "webpack": {
+        "version": "5.66.0",
+        "type": "--save-dev"
+      }
     },
     {
-      'webpack-bundle-analyzer': {
-        version: '4.5.0',
-        type: '--save-dev',
-      },
+      "webpack-bundle-analyzer": {
+        "version": "4.5.0",
+        "type": "--save-dev"
+      }
     },
     {
-      'webpack-merge': {
-        version: '5.8.0',
-        type: '--save-dev',
-      },
+      "webpack-merge": {
+        "version": "5.8.0",
+        "type": "--save-dev"
+      }
     },
     {
-      'webpack-cli': {
-        version: '4.9.1',
-        type: '--save-dev',
-      },
+      "webpack-cli": {
+        "version": "4.9.1",
+        "type": "--save-dev"
+      }
     },
     {
-      'webpack-dev-server': {
-        version: '4.5.0',
-        type: '--save-dev',
-      },
+      "webpack-dev-server": {
+        "version": "4.7.3",
+        "type": "--save-dev"
+      }
     },
     {
-      'resolve-url-loader': {
-        version: '4.0.0',
-        type: '--save-dev',
-      },
+      "resolve-url-loader": {
+        "version": "4.0.0",
+        "type": "--save-dev"
+      }
     },
     {
-      '@mparulski/es-project-maker-webpack': {
-        version: '0.9.0',
-        type: '--save-dev',
-      },
-    },
-  ],
+      "@mparulski/es-project-maker-webpack": {
+        "version": "0.9.0",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }